- fix bug #1543889: "Return value for Beanshell snippet" problem BeanShell was not properly overwriting new methods added with the same signature of old method.
There were two side-effects of this:

 (i) the first method with a particular name and signature would always be used, forever, even if a new method with the same signature was added. This broke the way the SearchAndReplace class cached the BS snippet used to do the substitution.

(ii) Since the method was still "compiled" and added to the internal list, there would be a slow memory leak.

Fix is to properly compare the signature of the new method with any old method registered with the same name, and get rid of the old method if it matches.

git-svn-id: https://svn.code.sf.net/p/jedit/svn/jEdit/trunk@6809 6b1eeb88-9816-0410-afa2-b43733a0f04e